修复频道入口与分区内容加载
- 更新全站与原创排行榜接口及导航参数，移除已失效的排行周期和评分字段
- 修正活动中心地址、过滤无效频道，并迁移番剧索引接口与父分区内容回退逻辑
- 扩展音频数据标识类型，记录首页加载异常，并调整横幅翻页按钮避免遮挡

+++ BiliBili.UWP/Pages/Music/MusicInfoPage.xaml.cs
@@ -492,11 +492,11 @@ namespace BiliBili.UWP.Pages
         public List<MusicHomeSongModel> list { get; set; }
 
 
-        public int id { get; set; }
-        public int up_mid { get; set; }
+        public long id { get; set; }
+        public long up_mid { get; set; }
         public string title { get; set; }
         public int duration { get; set; }
-        public int mid { get; set; }
+        public long mid { get; set; }
         public string cover_url { get; set; }
         public string intro { get; set; }
         public string  ctime_str { get; set; }
